Fix bugs in `ComponentAccess`, start reorganization
