--- blah/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "blah"
-edition.workspace = true
-version.workspace = true
-license.workspace = true
-repository.workspace = true
-
-[dependencies]
-
-[lints]
-workspace = true
