+++ src/query.rs
@@ -1,5 +1,20 @@
 //! Type-level DSL for retrieving data from entities.
 
+mod xor;
+mod has;
+mod not;
+mod option;
+mod tuple;
+mod or;
+mod with;
+mod misc;
+
+pub use xor::*;
+pub use has::*;
+pub use not::*;
+pub use or::*;
+pub use with::*;
+
 use core::fmt;
 use core::marker::PhantomData;
 use core::ptr::NonNull;
@@ -12,7 +27,7 @@ use crate::archetype::{Archetype, ArchetypeRow};
 use crate::assert::{AssertMutable, UnwrapDebugChecked};
 use crate::component::{Component, ComponentIdx};
 use crate::entity::EntityId;
-use crate::handler::{Config, InitError};
+use crate::handler::{HandlerConfig, InitError};
 use crate::world::World;
 
 /// Types that can be fetched from an entity.
@@ -71,8 +86,8 @@ pub unsafe trait Query {
     /// accessed by the query and a new instance of [`Self::State`].
     fn init(
         world: &mut World,
-        config: &mut Config,
-    ) -> Result<(ArchetypeFilter, ComponentAccess, Self::State), InitError>;
+        config: &mut HandlerConfig,
+    ) -> Result<(ComponentAccess, Self::State), InitError>;
 
     /// Returns a new [`Self::State`] instance.
     fn new_state(world: &mut World) -> Self::State;
@@ -104,631 +119,6 @@ pub unsafe trait Query {
 ///
 /// [`Fetcher::get`]: crate::fetch::Fetcher::get
 pub unsafe trait ReadOnlyQuery: Query {}
-
-unsafe impl<C: Component> Query for &'_ C {
-    type Item<'a> = &'a C;
-
-    type ArchState = ColumnPtr<C>;
-
-    type State = ComponentIdx;
-
-    fn init(
-        world: &mut World,
-        config: &mut Config,
-    ) -> Result<(ArchetypeFilter, ComponentAccess, Self::State), InitError> {
-        let idx = Self::new_state(world);
-        let filter = ArchetypeFilter::var(idx);
-        let ca = ComponentAccess::var(idx, Access::Read);
-        config.referenced_components.insert(idx);
-
-        Ok((filter, ca, idx))
-    }
-
-    fn new_state(world: &mut World) -> Self::State {
-        world.add_component::<C>().index()
-    }
-
-    fn new_arch_state(arch: &Archetype, state: &mut Self::State) -> Option<Self::ArchState> {
-        arch.column_of(*state).map(|c| ColumnPtr(c.data().cast()))
-    }
-
-    unsafe fn get<'a>(state: &Self::ArchState, row: ArchetypeRow) -> Self::Item<'a> {
-        &*state.0.as_ptr().cast_const().add(row.0 as usize)
-    }
-}
-
-unsafe impl<C: Component> ReadOnlyQuery for &'_ C {}
-
-unsafe impl<C: Component> Query for &'_ mut C {
-    type Item<'a> = &'a mut C;
-
-    type ArchState = ColumnPtr<C>;
-
-    type State = ComponentIdx;
-
-    fn init(
-        world: &mut World,
-        config: &mut Config,
-    ) -> Result<(ArchetypeFilter, ComponentAccess, Self::State), InitError> {
-        let () = AssertMutable::<C>::COMPONENT;
-
-        let idx = Self::new_state(world);
-        let filter = ArchetypeFilter::var(idx);
-        let ca = ComponentAccess::var(idx, Access::ReadWrite);
-        config.referenced_components.insert(idx);
-
-        Ok((filter, ca, idx))
-    }
-
-    fn new_state(world: &mut World) -> Self::State {
-        world.add_component::<C>().index()
-    }
-
-    fn new_arch_state(arch: &Archetype, state: &mut Self::State) -> Option<Self::ArchState> {
-        <&C>::new_arch_state(arch, state)
-    }
-
-    unsafe fn get<'a>(state: &Self::ArchState, row: ArchetypeRow) -> Self::Item<'a> {
-        &mut *state.0.as_ptr().add(row.0 as usize)
-    }
-}
-
-macro_rules! impl_query_tuple {
-    ($(($Q:ident, $q:ident)),*) => {
-        #[allow(unused_variables, clippy::unused_unit)]
-        unsafe impl<$($Q: Query),*> Query for ($($Q,)*) {
-            type Item<'a> = ($($Q::Item<'a>,)*);
-
-            type ArchState = ($($Q::ArchState,)*);
-
-            type State = ($($Q::State,)*);
-
-            #[allow(unused_mut)]
-            fn init(
-                world: &mut World,
-                config: &mut Config
-            ) -> Result<(ArchetypeFilter, ComponentAccess, Self::State), InitError> {
-                let mut filter = ArchetypeFilter::from(true);
-                let mut ca = ComponentAccess::new();
-
-                $(
-                    let (this_filter, this_ca, $q) = $Q::init(world, config)?;
-
-                    filter = filter.and(&this_filter);
-                    ca = ca.and(&this_ca);
-                )*
-
-                Ok((filter, ca, ($($q,)*)))
-            }
-
-            fn new_state(world: &mut World) -> Self::State {
-                (
-                    $(
-                        $Q::new_state(world),
-                    )*
-                )
-            }
-
-            fn new_arch_state(arch: &Archetype, ($($q,)*): &mut Self::State) -> Option<Self::ArchState> {
-                Some((
-                    $(
-                        $Q::new_arch_state(arch, $q)?,
-                    )*
-                ))
-            }
-
-            unsafe fn get<'a>(($($q,)*): &Self::ArchState, row: ArchetypeRow) -> Self::Item<'a> {
-                (
-                    $(
-                        $Q::get($q, row),
-                    )*
-                )
-            }
-        }
-
-        unsafe impl<$($Q: ReadOnlyQuery),*> ReadOnlyQuery for ($($Q,)*) {}
-    }
-}
-
-// Currently, debug impls for tuples only go up to arity 12.
-all_tuples!(impl_query_tuple, 0, 12, Q, q);
-
-/// Returns the result of `Q` as `Some`, or `None` if `Q` does not match.
-unsafe impl<Q: Query> Query for Option<Q> {
-    type Item<'a> = Option<Q::Item<'a>>;
-
-    type ArchState = Option<Q::ArchState>;
-
-    type State = Q::State;
-
-    fn init(
-        world: &mut World,
-        config: &mut Config,
-    ) -> Result<(ArchetypeFilter, ComponentAccess, Self::State), InitError> {
-        let (_, ca, state) = Q::init(world, config)?;
-
-        // Access components as `Q` does, but without affecting the matched archetypes.
-        Ok((ArchetypeFilter::from(true), ca, state))
-    }
-
-    fn new_state(world: &mut World) -> Self::State {
-        Q::new_state(world)
-    }
-
-    fn new_arch_state(arch: &Archetype, state: &mut Self::State) -> Option<Self::ArchState> {
-        Some(Q::new_arch_state(arch, state))
-    }
-
-    unsafe fn get<'a>(state: &Self::ArchState, row: ArchetypeRow) -> Self::Item<'a> {
-        state.as_ref().map(|f| Q::get(f, row))
-    }
-}
-
-unsafe impl<Q: ReadOnlyQuery> ReadOnlyQuery for Option<Q> {}
-
-/// A [`Query`] which matches if the `L` or `R` queries match.
-#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
-pub enum Or<L, R> {
-    /// Only the left query matched.
-    Left(L),
-    /// Only the right query matched.
-    Right(R),
-    /// Both the left and right queries matched.
-    Both(L, R),
-}
-
-impl<L, R> Or<L, R> {
-    /// Convert `Or<L, R>` to `Or<M, S>` using the supplied closures.
-    pub fn map<F, G, M, S>(self, f: F, g: G) -> Or<M, S>
-    where
-        F: FnOnce(L) -> M,
-        G: FnOnce(R) -> S,
-    {
-        match self {
-            Or::Left(l) => Or::Left(f(l)),
-            Or::Right(r) => Or::Right(g(r)),
-            Or::Both(l, r) => Or::Both(f(l), g(r)),
-        }
-    }
-
-    /// Converts `&Or<L, R>` to `Or<&L, &R>`.
-    pub const fn as_ref(&self) -> Or<&L, &R> {
-        match self {
-            Or::Left(l) => Or::Left(l),
-            Or::Right(r) => Or::Right(r),
-            Or::Both(l, r) => Or::Both(l, r),
-        }
-    }
-
-    /// Converts `&mut Or<L, R>` to `Or<&mut L, &mut R>`.
-    pub fn as_mut(&mut self) -> Or<&mut L, &mut R> {
-        match self {
-            Or::Left(l) => Or::Left(l),
-            Or::Right(r) => Or::Right(r),
-            Or::Both(l, r) => Or::Both(l, r),
-        }
-    }
-}
-
-unsafe impl<L, R> Query for Or<L, R>
-where
-    L: Query,
-    R: Query,
-{
-    type Item<'a> = Or<L::Item<'a>, R::Item<'a>>;
-
-    type ArchState = Or<L::ArchState, R::ArchState>;
-
-    type State = (L::State, R::State);
-
-    fn init(
-        world: &mut World,
-        config: &mut Config,
-    ) -> Result<(ArchetypeFilter, ComponentAccess, Self::State), InitError> {
-        let (filter_lhs, ca_lhs, state_lhs) = L::init(world, config)?;
-        let (filter_rhs, ca_rhs, state_rhs) = R::init(world, config)?;
-
-        Ok((
-            filter_lhs.or(&filter_rhs),
-            ca_lhs.or(&ca_rhs),
-            (state_lhs, state_rhs),
-        ))
-    }
-
-    fn new_state(world: &mut World) -> Self::State {
-        (L::new_state(world), R::new_state(world))
-    }
-
-    fn new_arch_state(
-        arch: &Archetype,
-        (left_state, right_state): &mut Self::State,
-    ) -> Option<Self::ArchState> {
-        match (
-            L::new_arch_state(arch, left_state),
-            R::new_arch_state(arch, right_state),
-        ) {
-            (None, None) => None,
-            (None, Some(r)) => Some(Or::Right(r)),
-            (Some(l), None) => Some(Or::Left(l)),
-            (Some(l), Some(r)) => Some(Or::Both(l, r)),
-        }
-    }
-
-    unsafe fn get<'a>(state: &Self::ArchState, row: ArchetypeRow) -> Self::Item<'a> {
-        state.as_ref().map(|l| L::get(l, row), |r| R::get(r, row))
-    }
-}
-
-unsafe impl<L, R> ReadOnlyQuery for Or<L, R>
-where
-    L: ReadOnlyQuery,
-    R: ReadOnlyQuery,
-{
-}
-
-/// A [`Query`] which matches if the `L` or `R` queries match, but not both.
-#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
-pub enum Xor<L, R> {
-    /// Only the left query matched.
-    Left(L),
-    /// Only the right query matched.
-    Right(R),
-}
-
-impl<L, R> Xor<L, R> {
-    /// Convert `Xor<L, R>` to `Xor<M, S>` using the supplied closures.
-    pub fn map<F, G, M, S>(self, f: F, g: G) -> Xor<M, S>
-    where
-        F: FnOnce(L) -> M,
-        G: FnOnce(R) -> S,
-    {
-        match self {
-            Xor::Left(l) => Xor::Left(f(l)),
-            Xor::Right(r) => Xor::Right(g(r)),
-        }
-    }
-
-    /// Convert `&Xor<L, R>` to `Xor<&L, &R>`.
-    pub const fn as_ref(&self) -> Xor<&L, &R> {
-        match self {
-            Xor::Left(l) => Xor::Left(l),
-            Xor::Right(r) => Xor::Right(r),
-        }
-    }
-
-    /// Convert `&mut Xor<L, R>` to `Xor<&mut L, &mut R>`.
-    pub fn as_mut(&mut self) -> Xor<&mut L, &mut R> {
-        match self {
-            Xor::Left(l) => Xor::Left(l),
-            Xor::Right(r) => Xor::Right(r),
-        }
-    }
-}
-
-unsafe impl<L, R> Query for Xor<L, R>
-where
-    L: Query,
-    R: Query,
-{
-    type Item<'a> = Xor<L::Item<'a>, R::Item<'a>>;
-
-    type ArchState = Xor<L::ArchState, R::ArchState>;
-
-    type State = (L::State, R::State);
-
-    fn init(
-        world: &mut World,
-        config: &mut Config,
-    ) -> Result<(ArchetypeFilter, ComponentAccess, Self::State), InitError> {
-        let (filter_lhs, ca_lhs, state_lhs) = L::init(world, config)?;
-        let (filter_rhs, ca_rhs, state_rhs) = R::init(world, config)?;
-
-        Ok((
-            filter_lhs.xor(&filter_rhs),
-            ca_lhs.xor(&ca_rhs),
-            (state_lhs, state_rhs),
-        ))
-    }
-
-    fn new_state(world: &mut World) -> Self::State {
-        (L::new_state(world), R::new_state(world))
-    }
-
-    fn new_arch_state(
-        arch: &Archetype,
-        (left_state, right_state): &mut Self::State,
-    ) -> Option<Self::ArchState> {
-        match (
-            L::new_arch_state(arch, left_state),
-            R::new_arch_state(arch, right_state),
-        ) {
-            (None, None) => None,
-            (None, Some(r)) => Some(Xor::Right(r)),
-            (Some(l), None) => Some(Xor::Left(l)),
-            (Some(_), Some(_)) => None,
-        }
-    }
-
-    unsafe fn get<'a>(state: &Self::ArchState, row: ArchetypeRow) -> Self::Item<'a> {
-        state.as_ref().map(|l| L::get(l, row), |r| R::get(r, row))
-    }
-}
-
-unsafe impl<L, R> ReadOnlyQuery for Xor<L, R>
-where
-    L: ReadOnlyQuery,
-    R: ReadOnlyQuery,
-{
-}
-
-/// A [`Query`] which matches if query `Q` doesn't match.
-pub struct Not<Q>(PhantomData<fn() -> Q>);
-
-impl<Q> Not<Q> {
-    /// Create a new instance.
-    pub const fn new() -> Self {
-        Self(PhantomData)
-    }
-}
-
-impl<Q> Clone for Not<Q> {
-    fn clone(&self) -> Self {
-        *self
-    }
-}
-
-impl<Q> Copy for Not<Q> {}
-
-impl<Q> Default for Not<Q> {
-    fn default() -> Self {
-        Self(Default::default())
-    }
-}
-
-impl<Q> fmt::Debug for Not<Q> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_tuple("Not").finish()
-    }
-}
-
-unsafe impl<Q: Query> Query for Not<Q> {
-    type Item<'a> = Self;
-
-    type ArchState = ();
-
-    type State = Q::State;
-
-    fn init(
-        world: &mut World,
-        config: &mut Config,
-    ) -> Result<(ArchetypeFilter, ComponentAccess, Self::State), InitError> {
-        let (filter, ca, state) = Q::init(world, config)?;
-
-        Ok((filter.not(), ca.not(), state))
-    }
-
-    fn new_state(world: &mut World) -> Self::State {
-        Q::new_state(world)
-    }
-
-    fn new_arch_state(arch: &Archetype, state: &mut Self::State) -> Option<Self::ArchState> {
-        match Q::new_arch_state(arch, state) {
-            Some(_) => None,
-            None => Some(()),
-        }
-    }
-
-    unsafe fn get<'a>(_state: &Self::ArchState, _row: ArchetypeRow) -> Self::Item<'a> {
-        Not::new()
-    }
-}
-
-unsafe impl<Q: Query> ReadOnlyQuery for Not<Q> {}
-
-/// A [`Query`] which matches if query `Q` matches.
-///
-/// Unlike `Q` however, `With<Q>` does not provide access to the data returned
-/// by `Q`. This is useful for avoiding access conflicts.
-///
-/// Example: `&C` requires read-only access to component `C`, but `With<&C>`
-/// does not require access at all.
-pub struct With<Q>(PhantomData<fn() -> Q>);
-
-impl<Q> With<Q> {
-    /// Create a new instance.
-    pub const fn new() -> Self {
-        Self(PhantomData)
-    }
-}
-
-impl<Q> Clone for With<Q> {
-    fn clone(&self) -> Self {
-        *self
-    }
-}
-
-impl<Q> Copy for With<Q> {}
-
-impl<Q> Default for With<Q> {
-    fn default() -> Self {
-        Self(Default::default())
-    }
-}
-
-impl<Q> fmt::Debug for With<Q> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_tuple("With").finish()
-    }
-}
-
-unsafe impl<Q: Query> Query for With<Q> {
-    type Item<'a> = Self;
-
-    type ArchState = ();
-
-    type State = Q::State;
-
-    fn init(
-        world: &mut World,
-        config: &mut Config,
-    ) -> Result<(ArchetypeFilter, ComponentAccess, Self::State), InitError> {
-        let (filter, mut ca, state) = Q::init(world, config)?;
-
-        ca.clear_access();
-
-        Ok((filter, ca, state))
-    }
-
-    fn new_state(world: &mut World) -> Self::State {
-        Q::new_state(world)
-    }
-
-    fn new_arch_state(arch: &Archetype, state: &mut Self::State) -> Option<Self::ArchState> {
-        Q::new_arch_state(arch, state).map(|_| ())
-    }
-
-    unsafe fn get<'a>(_state: &Self::ArchState, _row: ArchetypeRow) -> Self::Item<'a> {
-        With::new()
-    }
-}
-
-unsafe impl<Q: Query> ReadOnlyQuery for With<Q> {}
-
-/// A [`Query`] which returns a boolean indicating whether the query `Q`
-/// matches.
-///
-/// Like [`With`], `Has` does not provide access to the data returned by `Q`.
-pub struct Has<Q> {
-    has: bool,
-    _marker: PhantomData<fn() -> Q>,
-}
-
-impl<Q> Has<Q> {
-    /// Creates a new instance wrapping `has`.
-    pub const fn new(has: bool) -> Self {
-        Self {
-            has,
-            _marker: PhantomData,
-        }
-    }
-
-    /// Extracts the inner boolean.
-    pub const fn get(self) -> bool {
-        self.has
-    }
-}
-
-impl<Q> Clone for Has<Q> {
-    fn clone(&self) -> Self {
-        *self
-    }
-}
-
-impl<Q> Copy for Has<Q> {}
-
-impl<Q> Default for Has<Q> {
-    fn default() -> Self {
-        Self {
-            has: false,
-            _marker: PhantomData,
-        }
-    }
-}
-
-impl<Q> fmt::Debug for Has<Q> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_tuple("Has").field(&self.has).finish()
-    }
-}
-
-unsafe impl<Q: Query> Query for Has<Q> {
-    type Item<'a> = Self;
-
-    type ArchState = bool;
-
-    type State = Q::State;
-
-    fn init(
-        world: &mut World,
-        config: &mut Config,
-    ) -> Result<(ArchetypeFilter, ComponentAccess, Self::State), InitError> {
-        let (_, _, state) = Q::init(world, config)?;
-
-        Ok((ArchetypeFilter::from(true), ComponentAccess::new(), state))
-    }
-
-    fn new_state(world: &mut World) -> Self::State {
-        Q::new_state(world)
-    }
-
-    fn new_arch_state(arch: &Archetype, state: &mut Self::State) -> Option<Self::ArchState> {
-        Some(Q::new_arch_state(arch, state).is_some())
-    }
-
-    unsafe fn get<'a>(state: &Self::ArchState, _row: ArchetypeRow) -> Self::Item<'a> {
-        Self::new(*state)
-    }
-}
-
-unsafe impl<Q: Query> ReadOnlyQuery for Has<Q> {}
-
-/// Returns the `EntityId` of the matched entity.
-unsafe impl Query for EntityId {
-    type Item<'a> = Self;
-
-    type ArchState = ColumnPtr<EntityId>;
-
-    type State = ();
-
-    fn init(
-        _world: &mut World,
-        _config: &mut Config,
-    ) -> Result<(ArchetypeFilter, ComponentAccess, Self::State), InitError> {
-        Ok((ArchetypeFilter::from(true), ComponentAccess::new(), ()))
-    }
-
-    fn new_state(_world: &mut World) -> Self::State {}
-
-    fn new_arch_state(arch: &Archetype, (): &mut Self::State) -> Option<Self::ArchState> {
-        Some(ColumnPtr(unsafe {
-            NonNull::new(arch.entity_ids().as_ptr().cast_mut()).unwrap_debug_checked()
-        }))
-    }
-
-    unsafe fn get<'a>(state: &Self::ArchState, row: ArchetypeRow) -> Self::Item<'a> {
-        *state.0.as_ptr().add(row.0 as usize)
-    }
-}
-
-unsafe impl ReadOnlyQuery for EntityId {}
-
-/// Like `()`, the `PhantomData<T>` query always succeeds.
-unsafe impl<T: ?Sized> Query for PhantomData<T> {
-    type Item<'a> = Self;
-
-    type ArchState = ();
-
-    type State = ();
-
-    fn init(
-        _world: &mut World,
-        _config: &mut Config,
-    ) -> Result<(ArchetypeFilter, ComponentAccess, Self::State), InitError> {
-        Ok((ArchetypeFilter::from(true), ComponentAccess::new(), ()))
-    }
-
-    fn new_state(_world: &mut World) -> Self::State {}
-
-    fn new_arch_state(_arch: &Archetype, _state: &mut Self::State) -> Option<Self::ArchState> {
-        Some(())
-    }
-
-    unsafe fn get<'a>(_state: &Self::ArchState, _row: ArchetypeRow) -> Self::Item<'a> {
-        Self
-    }
-}
-
-unsafe impl<T: ?Sized> ReadOnlyQuery for PhantomData<T> {}
 
 /// Transparent wrapper around a [`NonNull`]. This implements [`Send`] and
 /// [`Sync`] unconditionally.
